Implement spec test assert_return invoke

+++ src/include/defines.h
@@ -1,10 +1,10 @@
 #ifndef WASMVM_DEFINES_DEF
 #define WASMVM_DEFINES_DEF
 
-typedef signed int      i32_t;
-typedef signed long     i64_t;
-typedef unsigned int    u32_t;
-typedef unsigned long   u64_t;
+typedef signed int          i32_t;
+typedef signed long         i64_t;
+typedef unsigned int        u32_t;
+typedef unsigned long long  u64_t;
 typedef float   f32_t;
 typedef double  f64_t;
